+++ projects/packages/forms/src/modules/file-field/view.js
@@ -6,6 +6,9 @@ import { clearInputError } from '../../contact-form/js/form-errors.js';
 
 const NAMESPACE = 'jetpack/field-file';
 
+const ENTER = 13;
+const SPACE = 32;
+
 let uploadToken = null;
 let tokenExpiry = null;
 
@@ -86,6 +89,37 @@ const formatBytes = ( size, decimals = 2 ) => {
 	return `${ numberFormat.format( formattedSize ) } ${ sizes[ i ] }`;
 };
 
+const getFileIcon = file => {
+	const config = getConfig( NAMESPACE );
+	const fileType = file.type.split( '/' )[ 0 ];
+	const fileExtension = file.name.split( '.' ).pop().toLowerCase();
+
+	const iconMap = {
+		image: 'png',
+		video: 'mp4',
+		audio: 'mp3',
+		document: 'pdf',
+		application: 'txt',
+	};
+
+	const extensionMap = {
+		pdf: 'pdf',
+		doc: 'doc',
+		docx: 'doc',
+		txt: 'txt',
+		ppt: 'ppt',
+		pptx: 'ppt',
+		xls: 'xls',
+		xlsx: 'xls',
+		csv: 'xls',
+		zip: 'zip',
+		sql: 'sql',
+		cal: 'cal',
+	};
+	const iconName = extensionMap[ fileExtension ] || iconMap[ fileType ] || 'txt';
+	return 'url(' + config.iconsPath + iconName + '.svg)';
+};
+
 /**
  * Add the file to the context.
  *
@@ -119,20 +153,19 @@ const addFileToContext = file => {
 	}
 
 	const clientFileId = performance.now() + '-' + Math.random();
-
-	const fileUrl =
+	const hasImage =
 		[ 'image/gif', 'image/jpg', 'image/png', 'image/jpeg' ].includes( file.type ) &&
-		URL.createObjectURL
-			? 'url(' + URL.createObjectURL( file ) + ')'
-			: null;
-
+		URL.createObjectURL;
+	const fileUrl = hasImage ? 'url(' + URL.createObjectURL( file ) + ')' : getFileIcon( file );
 	context.files.push( {
 		name: file.name,
 		formattedSize: formatBytes( file.size, 2 ),
+		hasIcon: ! hasImage,
 		isUploaded: false,
 		hasError: !! error,
 		id: clientFileId,
-		url: fileUrl,
+		url: hasImage ? fileUrl : null,
+		mask: ! hasImage ? fileUrl : null,
 		error,
 	} );
 
@@ -233,6 +266,12 @@ const { state, actions } = store( NAMESPACE, {
 	},
 
 	actions: {
+		handleKeyDown: event => {
+			if ( event.keyCode === ENTER || event.keyCode === SPACE ) {
+				event.preventDefault();
+				actions.openFilePicker( event );
+			}
+		},
 		/**
 		 * Open the file picker dialog.
 		 */
@@ -383,7 +422,30 @@ const { state, actions } = store( NAMESPACE, {
 			// Remove the file from the context
 			context.files = context.files.filter( fileObject => fileObject.id !== clientFileId );
 		},
+
+		removeFileKeydown: event => {
+			if ( event.keyCode === ENTER || event.keyCode === SPACE ) {
+				event.preventDefault();
+				actions.removeFile( event );
+			}
+		},
 	},
 
-	callbacks: {},
+	callbacks: {
+		focusElement: function () {
+			const { ref } = getElement();
+			setTimeout( () => {
+				ref.focus( { focusVisible: true } );
+			}, 100 );
+
+			return withScope( function () {
+				const dropzone = ref
+					.closest( '.jetpack-form-file-field__container' )
+					.querySelector( '.jetpack-form-file-field__dropzone-inner' );
+				setTimeout( () => {
+					dropzone.focus( { focusVisible: true } );
+				}, 100 );
+			} );
+		},
+	},
 } );