+++ projects/packages/masterbar/src/admin-menu/class-admin-menu.php
@@ -7,6 +7,7 @@
 
 namespace Automattic\Jetpack\Masterbar;
 
+use Automattic\Jetpack\Admin_UI\Admin_Menu as Jetpack_Admin_UI_Admin;
 use Automattic\Jetpack\Assets;
 use Automattic\Jetpack\Assets\Logo;
 use Automattic\Jetpack\Redirect;
@@ -17,6 +18,14 @@ require_once __DIR__ . '/class-base-admin-menu.php';
  * Class Admin_Menu.
  */
 class Admin_Menu extends Base_Admin_Menu {
+
+	/**
+	 * Register the hooks.
+	 */
+	public function __construct() {
+		parent::__construct();
+		add_action( 'admin_menu', array( $this, 'register_nav_unification_jetpack_menus' ), 999 );
+	}
 
 	/**
 	 * Create the desired menu output.
@@ -398,6 +407,41 @@ class Admin_Menu extends Base_Admin_Menu {
 	}
 
 	/**
+	 * Register the nav unification submenu items using the Jetpack APIs.
+	 *
+	 * This is needed because there's a bug in WP Core that ignores the position argument.
+	 *
+	 * @see https://core.trac.wordpress.org/ticket/52035
+	 *
+	 * @return void
+	 */
+	public function register_nav_unification_jetpack_menus() {
+		Jetpack_Admin_UI_Admin::add_menu(
+			esc_attr__( 'Activity Log', 'jetpack-masterbar' ),
+			__( 'Activity Log', 'jetpack-masterbar' ),
+			'manage_options',
+			'https://wordpress.com/activity-log/' . $this->domain,
+			/**
+			 * Core should ideally document null for no-callback arg. https://core.trac.wordpress.org/ticket/52539.
+			 */
+			null,
+			2
+		);
+
+		Jetpack_Admin_UI_Admin::add_menu(
+			esc_attr__( 'Backup', 'jetpack-masterbar' ),
+			__( 'Backup', 'jetpack-masterbar' ),
+			'manage_options',
+			'https://wordpress.com/backup/' . $this->domain,
+			/**
+			 * Core should ideally document null for no-callback arg. https://core.trac.wordpress.org/ticket/52539.
+			 */
+			null,
+			3
+		);
+	}
+
+	/**
 	 * Create Jetpack menu.
 	 *
 	 * @param int  $position  Menu position.
@@ -415,10 +459,6 @@ class Admin_Menu extends Base_Admin_Menu {
 			// @phan-suppress-next-line PhanTypeMismatchArgumentProbablyReal -- Core should ideally document null for no-callback arg. https://core.trac.wordpress.org/ticket/52539.
 			add_menu_page( esc_attr__( 'Jetpack', 'jetpack-masterbar' ), __( 'Jetpack', 'jetpack-masterbar' ), 'manage_options', 'jetpack', null, $icon, $position );
 		}
-		// @phan-suppress-next-line PhanTypeMismatchArgumentProbablyReal -- Core should ideally document null for no-callback arg. https://core.trac.wordpress.org/ticket/52539.
-		add_submenu_page( 'jetpack', esc_attr__( 'Activity Log', 'jetpack-masterbar' ), __( 'Activity Log', 'jetpack-masterbar' ), 'manage_options', 'https://wordpress.com/activity-log/' . $this->domain, null, 2 );
-		// @phan-suppress-next-line PhanTypeMismatchArgumentProbablyReal -- Core should ideally document null for no-callback arg. https://core.trac.wordpress.org/ticket/52539.
-		add_submenu_page( 'jetpack', esc_attr__( 'Backup', 'jetpack-masterbar' ), __( 'Backup', 'jetpack-masterbar' ), 'manage_options', 'https://wordpress.com/backup/' . $this->domain, null, 3 );
 
 		if ( self::DEFAULT_VIEW === $this->get_preferred_view( 'jetpack' ) ) {
 			$this->hide_submenu_page( 'jetpack', 'jetpack#/settings' );