+++ projects/packages/jetpack-mu-wpcom/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.5.0] - 2025-06-06
+### Changed
+- Update `Edit homepage` banner design and copy. [#43643]
+
+### Removed
+- Launchpad: Remove "Select a design" step from build task list. [#43820]
+
+### Fixed
+- Prevent disallowed blocks from being used inside inner comments. [#43573]
+
 ## [6.4.0] - 2025-06-03
 ### Added
 - Add a site badge and the site plan to the admin bar. [#42499]
@@ -1581,6 +1591,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Testing initial package release.
 
+[6.5.0]: https://github.com/Automattic/jetpack-mu-wpcom/compare/v6.4.0...v6.5.0
 [6.4.0]: https://github.com/Automattic/jetpack-mu-wpcom/compare/v6.3.1...v6.4.0
 [6.3.1]: https://github.com/Automattic/jetpack-mu-wpcom/compare/v6.3.0...v6.3.1
 [6.3.0]: https://github.com/Automattic/jetpack-mu-wpcom/compare/v6.2.0...v6.3.0